fix: user id 0
