+++ frontend/src/components/ProtectedRoute.tsx
@@ -1,26 +1,18 @@
 "use client";
 
-import { usePathname, useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { getCookie } from "@/api/cookie/cookies";
+import { useState } from "react";
+import useTokenValidator from "@/hooks/useTokenValidator";
+import useMyProfileUpdator from "@/hooks/useMyProfileUpdator";
 
 export default function ProtectedRoute({ children }: { children: any }) {
-  const pathname = usePathname();
-  const router = useRouter();
   const [isClient, setIsClient] = useState(false);
 
-  useEffect(() => {
-    const token = getCookie("access_token");
-    setIsClient(true);
+  useTokenValidator({ setIsClient });
+  useMyProfileUpdator();
 
-    if (!token && pathname !== "/login" && pathname !== "/login/otp") {
-      router.push("/login");
-    }
-  }, [pathname, router]);
-
-  if (isClient) {
-    return children;
+  if (!isClient) {
+    return <div />;
   }
 
-  return <div />;
+  return children;
 }
